API

SpatialDependence.assignments — Method

assignments(mc::AbstractMapClassification)

Get the vector of classes indices for each observation.

assignments(mc)[i] is the index of the class to which the $i$-th observation is assigned.

SpatialDependence.bounds — Function

bounds(mc::AbstractMapClassification)

Get the lower and upper bounds of the classes.

bounds(mc)[1][k] is the lower bound of the $k$-class.

bounds(mc)[2][k] is the upper bound of the $k$-class.

SpatialDependence.cardinalities — Method

cardinalities(W::SpatialWeights)

Return a vector with the number of neighbors of each observation.

SpatialDependence.centroid — Function

centroid(P)

Calculate the centroid of a polygon or vector of polygons $P$.

SpatialDependence.dnearneigh — Method

dnearneigh(A; threshold)

Build a spatial weights object from a table A that constains a points geometry column using a distance threshold.

SpatialDependence.dnearneigh — Method

dnearneigh(X, Y; threshold)

Build a spatial weights object from a set of coordinates using a distance threshold.

SpatialDependence.dnearneigh — Method

dnearneigh(P; threshold)

Build a spatial weights object from a vector of points using a distance threshold.

SpatialDependence.geary — Method

geary(x, W)

Compute the Geary's c test of global spatial autocorrelation.

Optional Arguments

permutations=9999: number of permutations for the randomization test.

rng=default_rng(): random number generator for the randomization test.

SpatialDependence.islands — Method

islands (W::SpatialWeights)

Return a vector with the islands in the spatial weights object.

SpatialDependence.knearneigh — Method

knearneigh(A; k)

Build a spatial weights object from a table A that constains a points geometry column using $k$ nearest neighbors..

SpatialDependence.knearneigh — Method

knearneigh(X, Y; k)

Build a spatial weights object from a set of coordinates using k nearest neighbors.

SpatialDependence.knearneigh — Method

knearneigh(P; k)

Build a spatial weights object from a vector of points using $k$ nearest neighbors.

SpatialDependence.level — Function

levels(mc::AbstractMapClassification)

Get the levels of the classes.

levels(mc)[k] is the level of the $k$-class.

SpatialDependence.mapclassify — Function

mapclassify (mcr::AbstractMapClassificator, x::Vector)

Classify observations in variable x in classess using the criterion specified in mcr.

SpatialDependence.maplabels — Function

maplabels (mcAbstractMapClassification)

Get the labels of the classification mc.

Optional Arguments

digits=2: number of decimal digits.

sep=", ": separator between class lower and upper bounds.

counts=true: include the total number of observations on each class.

SpatialDependence.meancenter — Function

meancenter(P)

Calculate the mean center of a polygon or vector of polygons $P$.

SpatialDependence.moran — Method

moran(x, W)

Compute the Moran's I test of global spatial autocorrelation.

Optional Arguments

permutations=9999: number of permutations for the randomization test.

rng=default_rng(): random number generator for the randomization test.

SpatialDependence.neighbors — Method

neighbors(W::SpatialWeights, i::Int)

Return a vector with the neighbors of $i$.

SpatialDependence.nislands — Method

nislands (W::SpatialWeights)

Return the number of islands in the spatial weights object.

SpatialDependence.polyneigh — Method

polyneigh(A, criterion = :Queen)

Build a spatial weights object from table A that contains a geometry column.

Optional Arguments

criterion=:Queen: neighbour criterion. :Queen or :Rook.

tol=0.0: tolerance for polygon contiguity.

SpatialDependence.polyneigh — Method

polyneigh(P, criterion = :Queen)

Build a spatial weights object from a vector of polygons P.

Optional Arguments

criterion=:Queen: neighbour criterion. :Queen or :Rook.

tol=0.0: tolerance for polygon contiguity.

SpatialDependence.reggeomlattice — Method

reggeomlattice(n, m)

Create a regular geometry-lattice of $n$ rows and $m$ columns.

Optional Arguments

gtype=:polygon: geometry. :polygon or :point.

direction=:rightdown: direction. :rightup, :rightdown, :leftdown or :leftup.

SpatialDependence.slag — Method

slag(W, x)

Calculate the spatial lag of x using the spatial weights W.

SpatialDependence.wtransform! — Method

wtransform!(W::SpatialWeights, style::Symbol)

In-place transformation of the weights using the specified $style$.

Weights transformation

style can be one of the following:

:binary: 1 if neighbor, 0 if not.

:row: row-standardized. Each row sum equals one.

SpatialDependence.wtransform — Method

wtransform(W::SpatialWeights, style::Symbol)

Returns a transformed copy of the weights matrix.

SpatialDependence.wtransformation — Method

wtransformation (W::SpatialWeights)

Return the current transformation of the spatial weights.

StatsBase.counts — Method

counts(mc::AbstractMapClassification)

Get the vector of classess sizes.

counts(mc)[k] is the number of observations assigned to the $k$-class.

StatsBase.weights — Method

weights(W::SpatialWeights, i::Int)

Return a vector with the weights of the neighbors of $i$.