Spatial Weight Matrices

Polygon contiguity

SpatialDependence.polyneigh — Function

polyneigh(P, criterion = :Queen)

Build a spatial weights object from a vector of polygons P.

Optional Arguments

criterion=:Queen: neighbour criterion. :Queen or :Rook.

tol=0.0: tolerance for polygon contiguity.

Examples

julia> using SpatialDatasets

julia> guerry = sdataset("Guerry");

julia> W = polyneigh(guerry.geometry)
Spatial Weights object 
Observations: 85 
Transformation: binary
Average number of neighbors: 4.9412
Minimum nunmber of neighbors: 2
Maximum nunmber of neighbors: 8
Median number of neighbors: 5.0
Islands (isloated): 0
Density: 5.8131% 

Points distance threshold

SpatialDependence.dnearneigh — Function

dnearneigh(X, Y; threshold)

Build a spatial weights object from a set of coordinates using a distance threshold.

Examples

julia> using SpatialDatasets

julia> bostonhsg = sdataset("Bostonhsg");

julia> W = dnearneigh(bostonhsg.x, bostonhsg.y, threshold = 4)
Spatial Weights object 
Observations: 506 
Transformation: binary
Average number of neighbors: 72.2253
Minimum nunmber of neighbors: 1
Maximum nunmber of neighbors: 178
Median number of neighbors: 56.0
Islands (isloated): 0
Density: 14.2738% 

dnearneigh(P; threshold)

Build a spatial weights object from a vector of points using a distance threshold.

Examples

julia> using SpatialDatasets

julia> bostonhsg = sdataset("Bostonhsg");

julia> W = dnearneigh(bostonhsg.geometry, threshold = 4)
Spatial Weights object 
Observations: 506 
Transformation: binary
Average number of neighbors: 72.2253
Minimum nunmber of neighbors: 1
Maximum nunmber of neighbors: 178
Median number of neighbors: 56.0
Islands (isloated): 0
Density: 14.2738% 

Points K nearest neighbors

SpatialDependence.knearneigh — Function

knearneigh(X, Y; k)

Build a spatial weights object from a set of coordinates using k nearest neighbors.

Examples

julia> using SpatialDatasets

julia> bostonhsg = sdataset("Bostonhsg");

julia> W = knearneigh(bostonhsg.x, bostonhsg.y, k = 10)
Spatial Weights object 
Observations: 506 
Transformation: Binary
Average number of neighbors: 10.0
Minimum nunmber of neighbors: 10
Maximum nunmber of neighbors: 10
Median number of neighbors: 10.0
Islands (isloated): 0
Density: 1.9763% 

knearneigh(P; k)

Build a spatial weights object from a vector of points using k nearest neighbors.

Examples

julia> using SpatialDatasets

julia> bostonhsg = sdataset("Bostonhsg");

julia> W = knearneigh(bostonhsg.geometry, k = 10)
Spatial Weights object 
Observations: 506 
Transformation: Binary
Average number of neighbors: 10.0
Minimum nunmber of neighbors: 10
Maximum nunmber of neighbors: 10
Median number of neighbors: 10.0
Islands (isloated): 0
Density: 1.9763% 

Spatial Matrix transformation

SpatialDependence.wtransform! — Function

wtransform!(W::SpatialWeights, style::Symbol)

In-place transformation of the weights using the specified style.

Weights transformation

style can be one of the following:

:binary: 1 if neighbor, 0 if not.

:row: row-standardized. Each row sum equals one.

Examples

julia> W = SpatialWeights([0 1 0; 1 0 1; 0 1 0]);

julia> wtransform!(W, :row)

SpatialDependence.wtransform — Function

wtransform(W::SpatialWeights, style::Symbol)

Returns a transformed copy of the weights matrix.

SpatialDependence.wtransformation — Function

wtransformation (W::SpatialWeights)

Return the current transformation of the spatial weights.

Examples

julia> W = SpatialWeights([0 1 0; 1 0 1; 0 1 0]);

julia> wtransform!(W, :row)

julia> wtransformation(W)
:row

Spatial Matrix information

SpatialDependence.nislands — Function

nislands (W::SpatialWeights)

Return the number of islands in the spatial weights object.

Examples

julia> W = SpatialWeights([0 1 0; 1 0 1; 0 0 0]);

julia> nislands(W)
1

SpatialDependence.islands — Function

islands (W::SpatialWeights)

Return a vector with the islands in the spatial weights object.

Examples

julia> W = SpatialWeights([0 1 0; 1 0 1; 0 0 0]);

julia> islands(W)
1-element Vector{Int64}:
 3

SpatialDependence.neighbors — Function

neighbors(W::SpatialWeights, i::Int)

Return a vector with the neighbors of i.

Examples

julia> W = SpatialWeights([0 1 0; 1 0 1; 0 1 0]);

julia> neighbors(W, 2)
2-element Vector{Int64}:
 1
 3

StatsBase.weights — Function

weights(W::SpatialWeights, i::Int)

Return a vector with the weights of the neighbors of i.

Examples

julia> W = SpatialWeights([0 1 0; 1 0 1; 0 1 0]);

julia> wtransform!(W, :row)

julia> weights(W, 2)
2-element Vector{Float64}:
 0.5
 0.5